+++ allennlp/models/encoder_decoders/wikitables/wikitables_decoder_step.py
@@ -77,8 +77,8 @@ class WikiTablesDecoderStep(DecoderStep[WikiTablesDecoderState]):
         hidden_state, memory_cell = self._decoder_cell(decoder_input, (hidden_state, memory_cell))
 
         # (group_size, encoder_output_dim)
-        encoder_outputs = torch.stack([state.rnn_state[i].encoder_outputs for i in state.batch_indices])
-        encoder_output_mask = torch.stack([state.rnn_state[i].encoder_output_mask for i in state.batch_indices])
+        encoder_outputs = torch.stack([state.rnn_state[0].encoder_outputs[i] for i in state.batch_indices])
+        encoder_output_mask = torch.stack([state.rnn_state[0].encoder_output_mask[i] for i in state.batch_indices])
         attended_question, attention_weights = self.attend_on_question(hidden_state,
                                                                        encoder_outputs,
                                                                        encoder_output_mask)
